+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   * Replaced `DynamicMutableRealm.create` with `DynamicMutableRealm.copyToRealm` similar to `MutableRealm.copyToRealm`.
 
 ### Enhancements
-* None.
+* Support for embedded objects (Issue [#551](https://github.com/realm/realm-kotlin/issues/551))
 
 ### Fixed
 * Creating a `RealmConfiguration` off the main thread on Kotlin Native could crash with `IncorrectDereferenceException`. (Issue [#799](https://github.com/realm/realm-kotlin/issues/799))